裁床单逻辑 ， 工序工价

+++ jinpus-service/jinpus-tpms-api/src/main/java/com/jinpus/tpms/api/domain/WorkOrderDo.java
@@ -68,7 +68,7 @@ public class WorkOrderDo extends BaseDo  {
 	* 单位
 	*/
     @Schema(description="单位")
-    private Integer unit;
+    private Long unit;
 
 	/**
 	* 单价
